backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
numpy==1.21.0
cadquery==2.1
pytest==6.2.4
python-dotenv==0.19.0
gunicorn==20.1.0